Resplit US to match psp discovery
