+++ config/splat.us.dra.yaml
@@ -82,10 +82,10 @@ segments:
       - [0x2C0C, .data, 5087C] # D_800A2C0C, D_800A2D44
       - [0x2D48, .data, menu]
       - [0x2EC0, .data, 5D5BC] # g_LevelHPIncrease
-      - [0x2EE8, .data, 5D5BC] # hud sprites def
-      - [0x2FBC, .data, 5D5BC] # g_CapePaletteDefs
-      - [0x300C, .data, 5D5BC] # g_CheatCodes
-      - [0x3014, .data, 5D5BC] # D_800A3014
+      - [0x2EE8, .data, 5F60C] # hud sprites def
+      - [0x2FBC, .data, 5F60C] # g_CapePaletteDefs
+      - [0x300C, .data, 5F60C] # g_CheatCodes
+      - [0x3014, .data, 5F60C] # D_800A3014
       - [0x3134, .data, 628AC] # D_800A3194
       - [0x31B0, .data, 63ED4] # D_800A3210
       - [0x3330, .data, 63ED4] # D_800A33A0-D_800A35D0
@@ -176,6 +176,7 @@ segments:
       - [0x3CBAC, .rodata, 5D5BC] # func_800FD6C4
       # - [0x3CBC4, .rodata, 5D5BC] # func_800FD7C0
       # - [0x3CBDC, .rodata, 5D5BC] # GetStatusAilmentTimer
+      - [0x3CBF4, .rodata, 5F60C] # g_CheatCodes
       - [0x3CC0C, .rodata, 624DC] # func_801027C4
       - [0x3CC2C, .rodata, 628AC] # func_801028AC
       # - [0x3CC54, .rodata, 628AC] # func_80102D70
@@ -247,6 +248,7 @@ segments:
       - [0x5087C, c, 5087C]
       - [0x5483C, c, menu]
       - [0x5D5BC, c, 5D5BC]
+      - [0x5F60C, c, 5F60C]
       - [0x624DC, c, 624DC]
       - [0x628AC, c, 628AC]
       - [0x63ED4, c, 63ED4]
